Split Lean declaration tests by ownership
