--- tests/component/providers/lean/test_lean_declarations.py
@@ -1,551 +1,0 @@
-from __future__ import annotations
-
-import hashlib
-import threading
-from concurrent.futures import ThreadPoolExecutor
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any
-
-import pytest
-
-from jacobian.artifacts import ArtifactService
-from jacobian.builtin_capabilities import (
-    LeanDeclarationInspectAdapter,
-    LeanDeclarationSearchAdapter,
-    LeanDependencyGraphAdapter,
-)
-from jacobian.capability_service import CapabilityInvocationError
-from jacobian.contracts.capabilities import (
-    CapabilityAssuranceLevel,
-    CapabilityCompletenessStatus,
-    CapabilityInstallTier,
-    CapabilityMode,
-    CapabilityProviderAvailability,
-    CapabilityProviderDigestKind,
-    CapabilityProviderRuntime,
-    CapabilityRequest,
-)
-from jacobian.contracts.lean import LeanDependencyGraphArtifact, LeanEnvironment
-from jacobian.lean_frontend.declarations import (
-    LeanDeclarationBackend,
-    LeanDeclarationBackendError,
-    LeanDeclarationService,
-    LeanSubprocessDeclarationBackend,
-    _parse_session_response,
-)
-from jacobian.schema_registry import SchemaRegistry
-from jacobian.storage.repository import ArtifactRepository
-
-_DIGEST = "sha256:" + "a" * 64
-_RUNTIME = CapabilityProviderRuntime(
-    provider="jacobian.lean4",
-    availability=CapabilityProviderAvailability.AVAILABLE,
-    version="4.31.0",
-    digest="sha256:" + "b" * 64,
-    digest_kind=CapabilityProviderDigestKind.EXECUTABLE,
-    platform="test",
-    install_tier=CapabilityInstallTier.T3,
-    license_id="Apache-2.0",
-    features=("CORE", "MATHLIB"),
-)
-
-
-@dataclass
-class FakeBackend(LeanDeclarationBackend):
-    response: dict[str, Any]
-    calls: list[tuple[LeanEnvironment, dict[str, Any]]] = field(default_factory=list)
-
-    def environment_digest(self, _environment: LeanEnvironment) -> str:
-        return _DIGEST
-
-    def query(
-        self,
-        environment: LeanEnvironment,
-        payload: dict[str, Any],
-    ) -> dict[str, Any]:
-        self.calls.append((environment, payload))
-        return self.response
-
-
-class MissingDeclarationBackend:
-    def environment_digest(self, _environment: LeanEnvironment) -> str:
-        return _DIGEST
-
-    def query(
-        self,
-        _environment: LeanEnvironment,
-        _payload: dict[str, Any],
-    ) -> dict[str, Any]:
-        raise LeanDeclarationBackendError(
-            "LEAN_DECLARATION_NOT_FOUND",
-            "Lean did not find the exact declaration 'Missing.name'.",
-        )
-
-
-@dataclass
-class RecordingSession:
-    responses: list[dict[str, Any] | LeanDeclarationBackendError]
-    requests: list[dict[str, Any]] = field(default_factory=list)
-    closed: bool = False
-    active: int = 0
-    max_active: int = 0
-    request_started: threading.Event | None = None
-    release_request: threading.Event | None = None
-    after_request: Any = None
-    activity_lock: threading.Lock = field(default_factory=threading.Lock)
-
-    def request(
-        self,
-        payload: dict[str, Any],
-        *,
-        timeout_seconds: int,
-    ) -> dict[str, Any]:
-        del timeout_seconds
-        with self.activity_lock:
-            self.active += 1
-            self.max_active = max(self.max_active, self.active)
-        try:
-            self.requests.append(payload)
-            if self.request_started is not None:
-                self.request_started.set()
-            if self.release_request is not None:
-                assert self.release_request.wait(timeout=2)
-            response = self.responses.pop(0)
-            if isinstance(response, LeanDeclarationBackendError):
-                raise response
-            return response
-        finally:
-            if self.after_request is not None:
-                self.after_request()
-            with self.activity_lock:
-                self.active -= 1
-
-    def close(self) -> None:
-        self.closed = True
-
-
-class RecordingSubprocessBackend(LeanSubprocessDeclarationBackend):
-    def __init__(
-        self,
-        *,
-        lean_executable: Path,
-        provider_runtime: CapabilityProviderRuntime,
-        sessions: list[RecordingSession],
-    ) -> None:
-        super().__init__(
-            lean_executable=lean_executable,
-            mathlib_runtime=None,
-            provider_runtime=provider_runtime,
-        )
-        self.pending_sessions = sessions
-        self.started_sessions: list[RecordingSession] = []
-
-    def _start_session(
-        self,
-        environment: LeanEnvironment,
-        environment_digest: str,
-    ) -> RecordingSession:
-        del environment, environment_digest
-        session = self.pending_sessions.pop(0)
-        self.started_sessions.append(session)
-        return session
-
-
-def _recording_backend(
-    tmp_path: Path,
-    sessions: list[RecordingSession],
-) -> tuple[RecordingSubprocessBackend, Path]:
-    executable = tmp_path / "lean"
-    executable.write_bytes(b"pinned lean executable")
-    digest = "sha256:" + hashlib.sha256(executable.read_bytes()).hexdigest()
-    runtime = _RUNTIME.model_copy(update={"digest": digest, "features": ("CORE",)})
-    return (
-        RecordingSubprocessBackend(
-            lean_executable=executable,
-            provider_runtime=runtime,
-            sessions=sessions,
-        ),
-        executable,
-    )
-
-
-def test_search_adapter_exposes_bounded_computed_retrieval() -> None:
-    backend = FakeBackend(
-        {
-            "operation": "search",
-            "declarations": [
-                {
-                    "name": "irrational_sqrt_two",
-                    "type": "Irrational √2",
-                    "kind": "THEOREM",
-                    "namespace": None,
-                    "docstring": None,
-                    "source": {
-                        "module": "Mathlib.NumberTheory.Real.Irrational",
-                        "line": 143,
-                        "column": 8,
-                        "end_line": 143,
-                        "end_column": 27,
-                    },
-                    "match_reasons": ["NAME_SUBSTRING"],
-                }
-            ],
-            "scanned_declarations": 20_001,
-            "stop_reason": "RESULT_LIMIT",
-        }
-    )
-    adapter = LeanDeclarationSearchAdapter(
-        LeanDeclarationService(backend),
-        _RUNTIME,
-    )
-
-    result = adapter.invoke(
-        CapabilityRequest(
-            capability_id="lean.declaration.search",
-            mode=CapabilityMode.EXPLORE,
-            input={
-                "environment": "MATHLIB",
-                "name_contains": "irrational_sqrt_two",
-                "result_limit": 1,
-            },
-        )
-    )
-
-    assert result.assurance.level is CapabilityAssuranceLevel.COMPUTED
-    assert result.completeness.status is CapabilityCompletenessStatus.PARTIAL
-    assert result.output["environment_digest"] == _DIGEST
-    assert result.output["declarations"][0]["name"] == "irrational_sqrt_two"
-    assert result.scope is not None
-    assert result.scope.parameters["matching"] == (
-        "case-sensitive name substring and exact constants occurring in the "
-        "elaborated type"
-    )
-    assert backend.calls[0][1] == {
-        "operation": "search",
-        "declaration_name": None,
-        "name_contains": "irrational_sqrt_two",
-        "type_constants": [],
-        "namespace_prefixes": [],
-        "target_module_prefixes": [],
-        "kinds": [],
-        "limit": 1,
-        "max_depth": 0,
-        "max_nodes": 1,
-    }
-
-
-def test_exhausted_search_reports_computed_complete_coverage() -> None:
-    backend = FakeBackend(
-        {
-            "operation": "search",
-            "declarations": [],
-            "scanned_declarations": 626_944,
-            "stop_reason": "EXHAUSTED",
-        }
-    )
-    adapter = LeanDeclarationSearchAdapter(
-        LeanDeclarationService(backend),
-        _RUNTIME,
-    )
-
-    result = adapter.invoke(
-        CapabilityRequest(
-            capability_id="lean.declaration.search",
-            mode=CapabilityMode.EXPLORE,
-            input={
-                "environment": "MATHLIB",
-                "type_pattern": {"constants": ["Jacobian.DoesNotExist"]},
-            },
-        )
-    )
-
-    assert result.completeness.status is CapabilityCompletenessStatus.COMPLETE
-    assert result.completeness.assurance_level is CapabilityAssuranceLevel.COMPUTED
-    assert result.assurance.level is CapabilityAssuranceLevel.COMPUTED
-
-
-def test_inspect_adapter_returns_docs_without_promoting_the_theorem() -> None:
-    backend = FakeBackend(
-        {
-            "operation": "inspect",
-            "declaration": {
-                "name": "Nat.add",
-                "type": "Nat → Nat → Nat",
-                "kind": "DEFINITION",
-                "namespace": "Nat",
-                "docstring": "Addition of natural numbers.",
-                "source": None,
-                "match_reasons": [],
-            },
-        }
-    )
-    adapter = LeanDeclarationInspectAdapter(
-        LeanDeclarationService(backend),
-        _RUNTIME,
-    )
-
-    result = adapter.invoke(
-        CapabilityRequest(
-            capability_id="lean.declaration.inspect",
-            mode=CapabilityMode.EXPLORE,
-            input={
-                "environment": "CORE",
-                "declaration_name": "Nat.add",
-            },
-        )
-    )
-
-    assert result.assurance.level is CapabilityAssuranceLevel.COMPUTED
-    assert result.assurance.verification_record_uri is None
-    assert result.output["declaration"]["docstring"].startswith("Addition")
-    assert result.output["environment_digest"] == _DIGEST
-
-
-def test_dependency_adapter_exposes_partial_typed_subgraph(tmp_path: Path) -> None:
-    backend = FakeBackend(
-        {
-            "operation": "dependencies",
-            "nodes": [
-                {"name": "Nat.add_assoc", "kind": "THEOREM", "depth": 0},
-                {"name": "Nat.add", "kind": "DEFINITION", "depth": 1},
-            ],
-            "edges": [
-                {
-                    "source": "Nat.add_assoc",
-                    "target": "Nat.add",
-                    "kinds": ["TYPE", "VALUE"],
-                }
-            ],
-            "frontier": ["Nat.add"],
-            "node_budget_exhausted": False,
-            "closure_complete": False,
-        }
-    )
-    store = ArtifactRepository(tmp_path)
-    schemas = SchemaRegistry(store)
-    artifacts = ArtifactService(store, schemas)
-    semantics_uri = store.register_descriptor(
-        kind="semantics",
-        name="test.lean-dependencies",
-        version="1",
-        definition={"dependency_api": "Lean.Expr.getUsedConstantsAsSet"},
-    )
-    schema_uri = schemas.register(
-        name="test.lean-dependency-graph",
-        version="1",
-        schema=LeanDependencyGraphArtifact.model_json_schema(),
-    )
-    adapter = LeanDependencyGraphAdapter(
-        LeanDeclarationService(backend),
-        _RUNTIME,
-        artifacts,
-        semantics_uri=semantics_uri,
-        dependency_graph_schema_uri=schema_uri,
-    )
-
-    result = adapter.invoke(
-        CapabilityRequest(
-            capability_id="lean.declaration.dependencies",
-            mode=CapabilityMode.EXPLORE,
-            input={
-                "environment": "CORE",
-                "root_declaration": "Nat.add_assoc",
-                "max_depth": 1,
-                "max_nodes": 20,
-            },
-        )
-    )
-
-    assert result.assurance.level is CapabilityAssuranceLevel.COMPUTED
-    assert result.completeness.status is CapabilityCompletenessStatus.PARTIAL
-    assert result.output["edges"][0]["kinds"] == ["TYPE", "VALUE"]
-    assert result.output["closure_complete"] is False
-    assert result.output["dependency_graph_uri"] in result.artifact_uris
-    assert backend.calls[0][1]["operation"] == "dependencies"
-    assert backend.calls[0][1]["max_depth"] == 1
-    assert backend.calls[0][1]["max_nodes"] == 20
-
-
-def test_missing_declaration_is_an_explicit_failed_operation() -> None:
-    adapter = LeanDeclarationInspectAdapter(
-        LeanDeclarationService(MissingDeclarationBackend()),
-        _RUNTIME,
-    )
-
-    with pytest.raises(CapabilityInvocationError) as raised:
-        adapter.invoke(
-            CapabilityRequest(
-                capability_id="lean.declaration.inspect",
-                mode=CapabilityMode.EXPLORE,
-                input={
-                    "environment": "CORE",
-                    "declaration_name": "Missing.name",
-                },
-            )
-        )
-
-    assert raised.value.diagnostic.code == "LEAN_DECLARATION_NOT_FOUND"
-
-
-def test_environment_identity_fails_closed_if_the_lean_executable_changes(
-    tmp_path: Path,
-) -> None:
-    executable = tmp_path / "lean"
-    executable.write_bytes(b"pinned lean executable")
-    digest = "sha256:" + hashlib.sha256(executable.read_bytes()).hexdigest()
-    runtime = _RUNTIME.model_copy(update={"digest": digest, "features": ("CORE",)})
-    backend = LeanSubprocessDeclarationBackend(
-        lean_executable=executable,
-        mathlib_runtime=None,
-        provider_runtime=runtime,
-    )
-
-    assert backend.environment_digest(LeanEnvironment.CORE).startswith("sha256:")
-
-    executable.write_bytes(b"changed lean executable")
-
-    with pytest.raises(LeanDeclarationBackendError) as raised:
-        backend.environment_digest(LeanEnvironment.CORE)
-
-    assert raised.value.code == "LEAN_ENVIRONMENT_CHANGED"
-
-
-def test_subprocess_backend_reuses_one_pinned_session_per_environment(
-    tmp_path: Path,
-) -> None:
-    session = RecordingSession(
-        responses=[
-            {"operation": "search", "declarations": []},
-            {"operation": "inspect", "declaration": {"name": "Nat.add"}},
-        ]
-    )
-    backend, _ = _recording_backend(tmp_path, [session])
-
-    searched = backend.query(
-        LeanEnvironment.CORE,
-        {"operation": "search"},
-    )
-    inspected = backend.query(
-        LeanEnvironment.CORE,
-        {"operation": "inspect"},
-    )
-
-    assert len(backend.started_sessions) == 1
-    assert [request["operation"] for request in session.requests] == [
-        "search",
-        "inspect",
-    ]
-    assert searched["_environment_digest"] == inspected["_environment_digest"]
-    backend.close()
-    assert session.closed
-
-
-def test_subprocess_backend_serializes_concurrent_session_requests(
-    tmp_path: Path,
-) -> None:
-    request_started = threading.Event()
-    release_request = threading.Event()
-    session = RecordingSession(
-        responses=[
-            {"operation": "search", "declarations": []},
-            {"operation": "search", "declarations": []},
-        ],
-        request_started=request_started,
-        release_request=release_request,
-    )
-    backend, _ = _recording_backend(tmp_path, [session])
-
-    with ThreadPoolExecutor(max_workers=2) as executor:
-        first = executor.submit(
-            backend.query,
-            LeanEnvironment.CORE,
-            {"operation": "search"},
-        )
-        assert request_started.wait(timeout=1)
-        second_started = threading.Event()
-
-        def query_second() -> dict[str, Any]:
-            second_started.set()
-            return backend.query(
-                LeanEnvironment.CORE,
-                {"operation": "search"},
-            )
-
-        second = executor.submit(query_second)
-        assert second_started.wait(timeout=1)
-        release_request.set()
-        results = (
-            first.result(timeout=2),
-            second.result(timeout=2),
-        )
-
-    assert len(results) == 2
-    assert len(backend.started_sessions) == 1
-    assert session.max_active == 1
-    backend.close()
-
-
-def test_subprocess_backend_rejects_result_if_environment_changes_during_query(
-    tmp_path: Path,
-) -> None:
-    session = RecordingSession(responses=[{"operation": "search"}])
-    backend, executable = _recording_backend(tmp_path, [session])
-    session.after_request = lambda: executable.write_bytes(b"changed executable")
-
-    with pytest.raises(LeanDeclarationBackendError) as raised:
-        backend.query(LeanEnvironment.CORE, {"operation": "search"})
-
-    assert raised.value.code == "LEAN_ENVIRONMENT_CHANGED"
-    assert session.closed
-
-
-def test_subprocess_backend_discards_timed_out_session_before_retry(
-    tmp_path: Path,
-) -> None:
-    timed_out = RecordingSession(
-        responses=[
-            LeanDeclarationBackendError(
-                "LEAN_QUERY_TIMEOUT",
-                "Lean declaration discovery timed out.",
-            )
-        ]
-    )
-    replacement = RecordingSession(responses=[{"operation": "search"}])
-    backend, _ = _recording_backend(tmp_path, [timed_out, replacement])
-
-    with pytest.raises(LeanDeclarationBackendError) as raised:
-        backend.query(LeanEnvironment.CORE, {"operation": "search"})
-    result = backend.query(LeanEnvironment.CORE, {"operation": "search"})
-
-    assert raised.value.code == "LEAN_QUERY_TIMEOUT"
-    assert timed_out.closed
-    assert len(backend.started_sessions) == 2
-    assert result["operation"] == "search"
-    backend.close()
-
-
-def test_session_protocol_rejects_mismatched_request_identity() -> None:
-    line = (
-        'JACOBIAN_DECLARATION_RESULT {"request_id":"stale",'
-        '"payload":{"operation":"search"}}'
-    )
-
-    with pytest.raises(LeanDeclarationBackendError) as raised:
-        _parse_session_response(line, expected_request_id="current")
-
-    assert raised.value.code == "LEAN_QUERY_PROTOCOL_ERROR"
-
-
-def test_session_protocol_preserves_exact_missing_declaration_failure() -> None:
-    line = (
-        'JACOBIAN_DECLARATION_ERROR {"request_id":"current",'
-        '"code":"LEAN_DECLARATION_NOT_FOUND",'
-        '"message":"declaration not found: Missing.name"}'
-    )
-
-    with pytest.raises(LeanDeclarationBackendError) as raised:
-        _parse_session_response(line, expected_request_id="current")
-
-    assert raised.value.code == "LEAN_DECLARATION_NOT_FOUND"
-    assert raised.value.message == "Lean did not find the exact requested declaration."
